% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{row_suffle_variability}
\alias{row_suffle_variability}
\title{Determine the variability of matrices under row shuffling}
\usage{
row_suffle_variability(mat_lst, B = 50)
}
\arguments{
\item{mat_lst}{A list of matrices}

\item{B}{integer, Number of times to sample (shuffle)}
}
\value{
[B,K] matrix, Frobenius norm vectors corresponding to row shuffling
}
\description{
Determine the variability of matrices under row shuffling
}
