% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_tools.R
\name{vecnorm}
\alias{vecnorm}
\title{Compute Euclidean norm of vector}
\usage{
vecnorm(x)
}
\arguments{
\item{x}{[1,m] double, A vector of data}
}
\value{
[1,1] double, Euclidean norm of x
}
\description{
Convenience function for use with e.g. lapply
}
